requirements.txt
flask
openai>=1.3.8
python-dotenv
langchain
chromadb
sentence-transformers
beautifulsoup4
requests
lxml
playwright
langchain-community
unstructured